requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-extensions==3.2.1
drf-yasg==1.21.4
django-environ==0.9.0
